+++ config/initializers/constants/short_code_regex.rb
@@ -1,1 +1,1 @@
-SHORT_CODE_REGEX = %r{ ^[0-9a-zA-Z_]{6}$ }xi
+SHORT_CODE_REGEX = %r{ \A[0-9a-zA-Z_]{6}\z }x